agent_project/requirements.txt
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0
autogen-agentchat>=0.4.0
autogen-ext[local-code-execution]>=0.4.0
python-dotenv>=1.0.0 